+++ ice/CHANGELOG.md
@@ -1,7 +1,8 @@
 # webrtc-ice changelog
 
 ## Unreleased
-* Add IP filter to ICE `AgentConfig` [#306](https://github.com/webrtc-rs/webrtc/pull/306)
+
+* Add IP filter to ICE `AgentConfig` [#306](https://github.com/webrtc-rs/webrtc/pull/306) and [#318](https://github.com/webrtc-rs/webrtc/pull/318).
 
 ## v0.8.1